Fixes a concurrent access bug in Recent Files
I believe this may have been masked by either the old FoxTrot library
or the former Log4J configuration.  It looks like it has been around
for some time but remained hidden.
We're simply not going to do these accesses in Threads unless we find
it necessary.  The performance here should be fine without it.