Removing special CHIP handling
There are a number of special CHIP-handling features that get
activated by particular CHIP file names.  Most prominent is
GENE_SYMBOL, but there are a number of others as well.  This is
the start of removing all of this in favor of having CHIPs just
be transparently handled according to the basic file format.