Implemented final sort on Chip2chip; fixed Etiology bug
At the end of Chip2chip, sort the Gene Sets by name and also sort the
members.  This is for reproducibility of results.
Also, fix a bug in the Etiology tracking where all items were being
flagged as invalid for the target chip.  This was just an artifact of
the previous way of determining the mapping.