Remove forced capitalization of gene symbols
Modified the behavior so that gene symbols are left unchanged and are
processed just in the form in which were loaded.
This is to better support non-human species symbols (though actually
some types of human symbols are affected as well).  It's part of the
larger symbol conversion clean-up happening elsewhere (upcoming MSigDB
release, GSEA website).