Minor fixes
Fixes of issues found when trying to verify recent refactoring with the
test suite.  Also minor type edits, etc found during the process.
Fixed the "down" significance matrix level in the PermutationTest.
Reverted the mu & sigma adjustment fix in the stddev metric as this
made too large a change to the result to verify with the test suite. We
will bring this back later when a dedicated verification is possible.