Switch from XLS to TSV; Fix missing Symbol Mapping report for GSEA
This is Part I of the XLS->TSV switch; still need to clean up some
internals.
The XLS files were not real Excel files but just tab-delimited with
an XLS extension.  Excel has started to flag these as corrupt or
suspicious, so we're changing over to TSV.  Also switched from a
control character to a colon for intra-cell field separator, for better
readability and easier parsing.
Fixed GSEA to always produce the Symbol Mapping report when collapsing.