Preranked will annotate when Collapse/Remap is used
There was a bug in Preranked where it tried to look up annotations
in the Chip used for Collapse (invalid because the purpose of doing so
is obviously to change the namespace).  However, we have the necessary
info available from the Collapse process, so this change exposes that
for use in reporting.