Tweak for Jobs table visibility
We had a report of the Jobs table being unusable in a HiDPI setting
(apparently on Windows 10).  This tweak makes a bit more room available
for the table.  We could go a bit further here if necessary just trying
to make more room, but want to avoid a UI redesign.