File Chooser update for Mac notarization
Replace nearly all File Choosers with native widgets (AWT FileDialog)
to better support Mac notarization.  Apparently it's not possible to
select from reserved locations (Desktop, etc) when using a non-native
widget.  Since this is one of the main desired features it's necessary
even though a bit painful.
*Not* replacing the Directory Choosers on Linux & Windows as there's
no way to do this with FileDialog.  Fortunately, there's a hack to do
this on Mac.
This is the bulk of the work, but there's some clean-up remaining for
after initial platform testing.