Modified certain parsers to automatically handle NA and quoted values
Only modified the GCT and TXT parsers in this regard.  We decided not
to change the PCL and RES parsers since they are less frequently used
and are generally not created by humans.
Handles double-quoted values in the Name column and the data matrix,
to handle naive output from R, i.e. use of write.table without changing
the quoting default.
Handle NA in the data matrix as if it is a missing value.
Misc other improvements, like try-finally to enforce file close.