Changed a number of computations to ignore missing values
This is the behavior we claim in the documentation but it wasn't
happening in certain cases, including the Collapse modes and the
continuous Metrics.
I've reimplemented these to ignore NaN values (our 'marker' for
missing values) in the same way as the Vector.toVectorNaNless() code.
They should be a bit faster as well, though I haven't benchmarked them
yet.
This spills over in a couple of places not related to Collapse or
continuous Metrics, where I've implemented a second version of certain
computations to avoid the overhead of checking where it's known to be
safe.
This is just the first round of updates as we still need to follow
through these edits to make all callers & users handle missing values
as well.
Also removed a number of unused Metrics and refactored the rest to
simplify these objects.
Also fixed a minor bug in the Pearson code, where it makes absolute
integer comparisons against an FP Double value.  This was unnecessary,
so I've replaced those with something safe.