Refactored toward removal of unneeded functions
Much of this refactoring is to remove unnecessary method parameters
where they are being passed only a constant value, then following the
logic from there to simplify the method body based on the constant.
Also removed unnecessary call-through methods.
We are particularly trying to remove unnecessary math functions so
that is the starting point.  None of these math functions are being
removed in this branch; that will happen elsewhere.
Also, reformatted code for legibility, removed commented-out code, etc.