Added warnings for FDR issue with only one gene set, and NaN ES results
The FDR is not accurate when only one gene set is present, so we add a
warning to the report (note: post-filtering).
Added a check to the report to go through the EnrichmentResults and
look for NaN (or infinite) values in ES, NES, p-value, etc.  Log these
gene sets and then add a warning to the report.