Extended infinity-adjustment to all weighted modes; fixed lost warnings
The weighted scoring mode had an adjustment for infinite values; this
was extended to the other weighted modes (p1.5 and p2) and also to the
vector used for plotting the RankedList Profile in the main combo plot
of the report.
Found that warnings on datasets & ranked lists can get dropped during
certain transformations - collapse, de-duplication, filtering - since
we create a new object based on the transformed data matrix.  Fixed
this to make sure the warnings are forwarded or captured.