Improved infinity adjustments; fixed weighted_p1.5 bugs; Log4J2 update
Added a warning to the report for any infinite values coming out of the
metric calculations.
Changed the infinity adjustments for p2 and p1.5 weighted scoring modes
to match previous behavior but with better coverage.
Modified the infinity adjustment for plotting to match that of the
p2 and p1.5 modes.
Improved the infinity adjustments for better performance.
Switched to use Float.isFinite() where possible to avoid multiple tests
for isNaN and isInfinite (also Double.isFinite).
Fixed bugs in the p1.5 mode calculations.
Documented the KSCore optimization from Justin Guinney.
Updated to Log4J 2.16.0 which include Log4Shell mitigations.