Modified the GMT & CHIP selectors to add Mouse DB selection. WIP
Added new Tabs to each for selecting Mouse files.
Modified version sorting to track per-tab latest instead of global.
Changed CHIP sort order so Mouse target files sort ahead of others.
Modified to have CHIP & FTPFile carry a Version object, with Species,
for more efficient sorting and validation checking. This will be
extended to GMT / GeneSets as appropriate.
Refactored to move Swing UI component construction out of Params and
into the Chooser Windows, which is closer responsibility-wise.
Likewise, moved UI building out of the FTP data retrieval component
and into Chooser Windows.  Made the FTP connection reusable so we
can query for both Human and Mouse tabs with one connection. Also
made some efficiency improvements to the FTPFile object.
Changed the Chip Name comparator to work with FTPFiles so they can
efficiently work with the new Version object instead of always file
parsing.  This should make them usable for GMTs as well, but that
has not been done yet.
Changed the CHIP parser to add Version when the file is retrieved
from FTP.
Added a stub Validator object for the CHIP cross-tab check.  This
has not yet been wired into the File Selector, and also needs to
Added support for extra Info buttons to the File Selector dialog
component.
be expanded to add an error message.
Changed the report's link builder to only use URLs out of the GMT
rather than attempting to build them when that field is blank.
Added an item to the Downloads menu for the Mouse CHIPs.  The URLs
here still need to change when the final location is set.
Added exception stack-traces to the logging properties.