Improved validation for mixed Version and Species checking
Added support for Version checking, with certain Versioned objects
carrying a Version object to help with comparisons.
Added support to pass Validator objects into the DialogDescriptor
object so that the owner of the req data can define the validation
rules and failure message rather than having to share that data
or responsibility across the GUI.
Added the concept of an Unknown Version as an explicit object as
opposed to just using null.  This allows us to know when certain
objects (i.e. GeneSets) come from the "same" Unknown Version (i.e.
through a parent GeneSetMatrix).
Added the FTP download checks to the GMT & Chip parsers.  Was formerly
in the ParserFactory but that class is already far to big and complex,
and this allows us to be more targeted in setting the version.
Refactored the Chip and Gene Set choosers to have common methods in a
helper class and to simplify both overall.
Created an initial version of the check for mixed Versions and Species
across the CHIP vs Gene Set selections.  This is not final, just WIP.
Dropped the old ChipNameComparator as it is now unused.
Added support to Errors to have direct access to the underlying String
messages.